pyproject.toml
[project]
name = "bw2-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
    "requests>=2.32.3",
]
